Strongly typed responses are an attempt to provide clear results from JSON/XML responses. The prototype in this instance is the response for AppCreate and AppInfo. AppDestroy is a different response type that simply returns Unit -- it is not applicable as part of this change. AppList will need its own List<App> response type that is initialized differently from App.java.
Note that AppCreate uses JSON and AppInfo uses XML. While the response type of App is the same for the user of the API, the data responded with is different. This is the reason for composing the App POJO with the Response object, rather than having App extend Response.

App.java wraps Response by calling its get(ResponseKey) methods and applying simple type transformations (e.g. Object to Integer). ResponseKey has been added for convenience and clarity -- response objects should use the keys rather than Strings. Strings are left in, though, for forwards compatability.

At this point in time, App and the various Response implementations are not in the right package. Either request should be named something else, e.g. protocol, or the responses should be separate.