Refactored all requests to use JSON and Strong Types
All requests that respond with JSON return a type from com.heroku.api.model, a Unit, or a POJO that models the response with some helpers (e.g. LogStreamResponse). The Response interface has been removed -- since all the responses are POJOs, the Response interface doesn't offer any useful abstraction.

Access to the raw data has been removed as part of the removal of the Response interface. To gain access to the raw data, the appropriate means is to override the Request.getResponse. We can re-introduce this later if necessary, but wrapping the Response seems to make usage in the large majority of cases less clear (or, at least, more cumbersome).

Note: Currently, this modification requires the api_v2_cleanup UserPass.