1 ServiceLoader implementation for Connection
Connection needs to be instantiated from HerokuAPI. In order to do so without
introducing dependencies into heroku-api, a ServiceLoader is used to determine
an implementation at runtime.

Currently, each Connection implementation has a META-INF.services file with the
exception of FinagleConnection. I don't have enough Scala understanding at this
point to spend time trying to figure out how to make it work with ServiceLoader.
ConnectionFactory.get(HerokuAPIConfig) loads all ConnectionProvider implementations
and returns the first non-null Connection returned from
ConnectionProvider.get(HerokuAPIConfig).

Connection is still available as a constructor for HerokuAPI, however a
HerokuAPIConfig is the preferred method. It may make sense to remove the
new HerokuAPI(Connection) constructor and add it back in the future if its needed.

Usage is now simplified for the majority of users:
    HerokuAPI api = new HerokuAPI(new HerokuAPIConfig().setApiKey("09u3rknacvlkjdf09u214"));