Flattened the user-facing API into HerokuAPI
All verbs are located in the HerokuAPI class now. The API no longer manages state for
an app. Instead, the app name must be passed as a parameter just as they might be passed
to a REST endpoint. To create a new app, a builder is used rather than combinations of
strings. This is because there are two parameters, Stack and app name, that could be used
interchangeably -- the builder clarifies to the user which parameters they are specifying.

All model classes are located alongside HerokuAPI in com.heroku.api to reduce the total
number of imports. This means a user can import com.heroku.api.* and get all the classes
they need to work with (with the exception of com.heroku.api.exception).

HerokuAPIConfig is now used for configuring HerokuAPI. The config is a builder that expects
a username/password or an apiKey. If the user specifies a username/password, the apiKey can
be retrieved from HerokuAPI.getApiKey(). It may make sense to further separate the un/pw
usage by providing a HerokuAPI.obtainApiKey(un,pw) static method.

Usage:
    import com.heroku.api.*
    ...
    HerokuAPI api = new HerokuAPI(new HerokuAPIConfig().setApiKey("..."));
    App app = api.createApp(new App().on(Stack.Cedar).named("myapp"));