Simplified connection parameters for HerokuAPI
The HerokuAPIConfig builder was removed in favor of simple method parameters. The static method,
HerokuAPI.obtainApiKey(username, password) provides users a way to get an API key, but also
makes it obvious that a username and password are not the primary way to interact with the API.
The primary way to interact with the API is with an API key. The API key is passed into a
HerokuAPI constructor and used to construct a connection.

To get an API key programmatically:
    String apiKey = HerokuAPI.obtainApiKey("user@heroku.com", "password");

The apiKey is expected to be stored by the user in whatever means they find appropriate, e.g.
an encrypted file on the end user's system.

To use the api:
    HerokuAPI api = new HerokuAPI("apiKey01982hkdf09u234j");
    App app = api.getApp("appname");