Create the runtime instance of the jaxbContext in XmlParser once statically.
These are threadsafe, and creating the same context multiple times is ineffecient and causes jaxb to generate throw and catch several exceptions.

Fixes #33

One thing to remeber here is that we need to add any classes with Jaxb annotations to the jaxbContext in XmlParser. But it is likely we wont add any more
and will actually remove them.